finished raycasting & player movement/hitbox

+++ cub3d2/srcs/ft_check_map2.c
@@ -6,7 +6,7 @@
 /*   By: mbaxmann <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/25 10:36:53 by mbaxmann          #+#    #+#             */
-/*   Updated: 2020/07/29 09:51:13 by mbaxmann         ###   ########.fr       */
+/*   Updated: 2020/08/15 09:21:48 by mbaxmann         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int		ft_check_down(char **map, int i, int j)
 {
-	if (i && j <= ft_strlen(map[i - 1]) && map[i - 1][j] == ' ')
+	if (i && j <= (int)ft_strlen(map[i - 1]) && map[i - 1][j] == ' ')
 		return (1);
-	else if (!i || j > ft_strlen(map[i - 1]))
+	else if (!i || j > (int)ft_strlen(map[i - 1]))
 		return (1);
 	else if (map[i - 1][j] && map[i - 1][j + 1] == ' ')
 		return (1);
@@ -27,10 +27,10 @@ int		ft_check_down(char **map, int i, int j)
 
 int		ft_check_up(char **map, int i, int j)
 {
-	if (map[i + 1][0] && j <= ft_strlen(map[i + 1])
+	if (map[i + 1][0] && j <= (int)ft_strlen(map[i + 1])
 		&& map[i + 1][j] == ' ')
 		return (1);
-	else if (!map[i + 1][0] || j > ft_strlen(map[i + 1]))
+	else if (!map[i + 1][0] || j > (int)ft_strlen(map[i + 1]))
 		return (1);
 	else if (map[i + 1][j] && map[i + 1][j + 1] == ' ')
 		return (1);
